add comment functionality

+++ backend/src/main/java/me/project/backend/exception/notFound/PostNotFoundException.java
@@ -1,7 +1,7 @@
 package me.project.backend.exception.notFound;
 
 public class PostNotFoundException extends ResourceNotFoundException {
-    public PostNotFoundException(int postId) {
+    public PostNotFoundException(long postId) {
         super("Post with id " + postId);
     }
 }
